backend/pyproject.toml
[project]
name = "monday-qbr-copilot"
version = "0.1.0"
description = "QBR Co-Pilot backend — FastAPI + LangGraph agent infrastructure"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "langchain",
    "langgraph",
    "langchain-openai",
    "openpyxl",
    "python-dotenv",
    "ag-ui-protocol",
    "ag-ui-langgraph",
    "markdown>=3.10.2",
    "weasyprint>=68.1",
    "python-multipart>=0.0.26",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
